+++ __init__.py
@@ -3,7 +3,4 @@ __author__ = 'Ian Housman'
 __email__ = 'ian.housman@gmail.com'
 
 #Version format yyyy.m.n
-__version__ = '2023.7.5'
-
-
-
+__version__ = '2023.8.3'